Add test for TIM2(x) reserved peripheral memory
Show that the TIM2(x) CCR1 register's offset is not meeting its expected
position. This is because of how we determine reserved memory in the RAL
generation script. A child commit ensures that this test passes.

Only applies to the STM32L0 family. Run the test like

  cargo test --features=stm32l0x0 --tests

Tested on Debian 11.