Ensure alignment when computing reserved registers
If the routine for computing reserved peripheral memory places at least
one u32 after a u16 register, the compiler could insert an extra u16 of
padding to align the u32 reservation. The extra padding is not
considered by the routine. This can result in incorrect register offsets
for all registers placed after the reservation.

This commit ensures that we only use larger-than-byte padding when the
address is aligned for that primitive. If the address isn't aligned, use
the next-largest primitive. Byte array reservations should be the safest
approach, but going that route changes all _reserved members of all
register blocks, making it harder to see what peripherals were affected
by this issue.

This commit only includes the RAL generation script and changes for the
TIM2(x) peripherals in the STM32L0 family. This peripheral seems to
demonstrate the issue; basing this on the reference manual study, not
hardware testing. I've not studied all changes incured by a run of
'make,' so there may be other instances of this issue. This commit is
enough to fix the failing test in the parent commit.